Add/Fix Security Issues

+++ app.py
@@ -4,7 +4,7 @@ import hmac
 import time
 import jinja2
 import webapp2
-from model import db, post_key, users_key, User, Post
+from models import db, post_key, users_key, User, Post
 
 
 secret = 'secret'
@@ -12,13 +12,17 @@ secret = 'secret'
 
 # Jinja initialization
 template_dir = os.path.join(os.path.dirname(__file__), 'templates')
-jinja_env = jinja2.Environment(loader = jinja2.FileSystemLoader(template_dir), autoescape = True)
+jinja_env = jinja2.Environment(
+    loader=jinja2.FileSystemLoader(template_dir), autoescape=True)
+
 
 def render_str(template, **params):
     t = jinja_env.get_template(template)
     return t.render(params)
 
+
 class BlogHandler(webapp2.RequestHandler):
+
     def write(self, *a, **kw):
         self.response.out.write(*a, **kw)
 
@@ -51,10 +55,21 @@ class BlogHandler(webapp2.RequestHandler):
         self.user = uid and User.by_id(int(uid))
 
 
+# Security Layer
+def user_owns_post(self, post):
+    if post:
+        return self.user.name == post.author.name
+
+def user_logged_in(self):
+    return self.user
+
+
+
 
 # Hashing Methods
 def make_secure_val(val):
     return '%s|%s' % (val, hmac.new(secret, val).hexdigest())
+
 
 def check_secure_val(secure_val):
     val = secure_val.split('|')[0]
@@ -64,20 +79,25 @@ def check_secure_val(secure_val):
 
 # Input verification methods
 USER_RE = re.compile(r"^[a-zA-Z0-9_-]{3,20}$")
+
+
 def valid_username(username):
     return username and USER_RE.match(username)
 
 PASS_RE = re.compile(r"^.{3,20}$")
+
+
 def valid_password(password):
     return password and PASS_RE.match(password)
 
 
 
-
 # Home Page (Login)
 class MainPage(BlogHandler):
+
     def get(self):
         self.render('mainPage.html')
+
     def post(self):
         username = self.request.get('username')
         password = self.request.get('password')
@@ -88,12 +108,12 @@ class MainPage(BlogHandler):
             self.redirect('/blog')
         else:
             msg = 'Invalid login'
-            self.render('mainPage.html', error = msg)
-
+            self.render('mainPage.html', error=msg)
 
 
 # Sign Up Page
 class SignUpPage(BlogHandler):
+
     def get(self):
         self.render("signUpPage.html")
 
@@ -107,13 +127,13 @@ class SignUpPage(BlogHandler):
 
         if not valid_username(self.username):
             error_message = "That's not a valid username."
-            self.render('signUpPage.html', error_message = error_message)
+            self.render('signUpPage.html', error_message=error_message)
         elif not valid_password(self.password):
             error_message = "That wasn't a valid password."
-            self.render('signUpPage.html', error_message = error_message)
+            self.render('signUpPage.html', error_message=error_message)
         elif self.password != self.verify:
             error_message = "Your passwords didn't match."
-            self.render('signUpPage.html', error_message = error_message)
+            self.render('signUpPage.html', error_message=error_message)
         else:
             self.done()
 
@@ -123,11 +143,11 @@ class SignUpPage(BlogHandler):
 
 class Register(SignUpPage):
     def done(self):
-        #make sure the user doesn't already exist
+        # make sure the user doesn't already exist
         u = User.by_name(self.username)
         if u:
             msg = 'That user already exists.'
-            self.render('signUpPage.html', error_message = msg)
+            self.render('signUpPage.html', error_message=msg)
         else:
             u = User.register(self.username, self.password)
             u.put()
@@ -135,39 +155,52 @@ class Register(SignUpPage):
             self.redirect('/blog')
 
 
+
+
 # Blog Page
-class BlogFront(Register):
+class BlogFront(BlogHandler):
+
     def get(self):
-        posts = Post.all().order('-created')
-        params = dict(username = self.user.name, posts = posts)
-        time.sleep(0.1)
-        self.render('blogPage.html', **params)
+        # checks if user is logged in before showing blog entries
+        if user_logged_in(self):
+            posts = Post.all().order('-created')
+            params = dict(username=self.user.name, posts=posts)
+            time.sleep(0.1)
+            self.render('blogPage.html', **params)
+        else:
+            self.redirect('/')
 
 
 
 # New Post Creation
 class NewPost(BlogHandler):
+
     def get(self):
-        if self.user:
-            self.render("newPostPage.html", username = self.user.name)
+        # checks if user is logged in
+        if not user_logged_in(self):
+            self.redirect('/')
         else:
-            self.redirect("/")
+            self.render("newPostPage.html", username=self.user.name)
 
     def post(self):
-        if not self.user:
-            self.redirect('/blog')
+        if not user_logged_in(self):
+            self.redirect('/')
 
         title = self.request.get('title')
         content = self.request.get('content')
 
         if title and content:
-            p = Post(parent = post_key(), title = title, content = content, author = self.user.name)
+            p = Post(parent=post_key(), title=title,
+                     content=content, author=self.user)
             p.put()
             self.redirect('/blog')
             # self.redirect('/blog/%s' % str(p.key().id()))
         else:
             error = "Both subject and content are required!"
-            self.render("newPostPage.html", title = title, content = content, error = error)
+            self.render("newPostPage.html", title=title,
+                        content=content, error=error)
+
+
 
 
 # Edit Post Page
@@ -175,21 +208,32 @@ class PostPage(BlogHandler):
     def get(self, post_id):
         key = db.Key.from_path('Post', int(post_id), parent=post_key())
         post = db.get(key)
-
         if not post:
-            self.error(404)
-            return
+            return self.error(404)
 
-        if post.author != self.user.name:
+        if not user_logged_in(self):
             self.redirect('/')
 
-        params = dict(title = post.title, content = post.content, username = self.user.name)
-
-        self.render('editPostPage.html', **params)
+        if user_owns_post(self, post):
+            params = dict(title=post.title, content=post.content,
+                      username=self.user.name)
+            self.render('editPostPage.html', **params)
+        else:
+            self.redirect('/')
 
     def post(self, post_id):
         key = db.Key.from_path('Post', int(post_id), parent=post_key())
         post = db.get(key)
+        if not post:
+            return self.error(404)
+
+        # checks if user is logged in
+        if not user_logged_in(self):
+            self.redirect('/')
+        # checks if the user owns the post
+        if not user_owns_post(self, post):
+            self.redirect('/')
+
         title = self.request.get('title')
         content = self.request.get('content')
 
@@ -200,22 +244,37 @@ class PostPage(BlogHandler):
             self.redirect('/blog')
         else:
             error = "Both subject and content are required!"
-            self.render("newPostPage.html", title = title, content = content, error = error)
+            self.render("newPostPage.html", title=title,
+                        content=content, error=error)
+
 
 
 # Logout
 class Logout(BlogHandler):
+
     def get(self):
         self.logout()
         self.redirect('/')
 
+
 # Delete Post
+
 class DeletePost(BlogHandler):
+
     def get(self, post_id):
-        key = db.Key.from_path('Post', int(post_id), parent=post_key())
-        post = db.get(key)
-        post.delete()
-        self.redirect('/blog')
+
+        if user_logged_in(self):
+            key = db.Key.from_path('Post', int(post_id), parent=post_key())
+            post = db.get(key)
+            if not post:
+                return self.error(404)
+
+            if user_owns_post(self, post):
+                post.delete()
+                self.redirect('/blog')
+        else:
+            self.redirect('/')
+
 
 
 app = webapp2.WSGIApplication([('/', MainPage),
@@ -228,3 +287,11 @@ app = webapp2.WSGIApplication([('/', MainPage),
                                ('/logout', Logout)
                                ],
                               debug=True)
+
+
+"""
+TODO:
+    - Add comments feature
+    - Add Like/Dislike feature
+
+"""
